// src/generated/registerPasskeyRequest.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * TeaCoder API
 * API for Teacoder educational platform
 * OpenAPI spec version: 1.0.0
 */

export interface RegisterPasskeyRequest {
  /**
   * Name of the device used for registering the passkey
   * @maxLength 50
   */
  deviceName: string;
  /**
   * Unique credential ID for the passkey
   * @maxLength 255
   */
  credentialId: string;
  /**
   * Public key for the passkey
   * @maxLength 2048
   */
  publicKey: string;
  /** List of supported transport methods for the passkey */
  transports: string[];
}